Make provenance of method name docs clearer
This adds a description to the link to the "metering primitives" page from
the Traffic Control project, which then makes the sentence flow better; a
link to the Traffic Control project has also been added, since other parts
of the documentation also point there and it seems to make sense to point to
the origin of the documentation.  The sentence flow has also been changed so
that it flows a bit better.