Point broken link to web.archive.org stored version
Unfortunately the assignments for cs143 at Harvard seem to require a login
to the Harvard website (the original link has been missing now for several
years).  The original text can now be accessed via the Internet Archive's
archived version; the version of the document linked here is the most recent
existing version.